requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
dataclasses==0.6
Django==5.0
django-bootstrap5==24.2
environs==11.0.0
h11==0.14.0
idna==3.7
marshmallow==3.21.1
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
pony==0.7.17
PySocks==1.7.1
python-dotenv==1.0.1
reportlab==4.2.0
requests==2.31.0
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
